Fix insertion of spurious U+FEFF in UTF-8-SIG files
Don't insert U+FEFF characters after each replacement if the file is
UTF-8 with a Byte Order Mark. Using BOM doesn't make much sense in UTF-8
files, yet such files exist and rpl would previously corrupt them by
inserting the mark after each replacement. Because the inserted U+FEFF
ZERO WIDTH NO-BREAK SPACE is an invisible character, this was easy to
miss.

This was caused by the way rpl encodes the output in pieces. Python's
UTF-8-SIG codec is stateful and only outputs BOM at the beginning, but
the way strings are encoded in rpl, "beginning" is after every match as
well.

Fixed by simply ignoring the special-case encoding and treating it as
UTF-8. This is safe to do because UTF-8 with BOM is valid UTF-8 too.