Fix a performance problem (fix #21)
Attempt to avoid adding the entire input to a string, one buffer-full at a
time, by determining the maximum match length of match for the given regex,
and skipping over input that is more than the max. length of the match
before the end of the buffer.

Patterns like ‘a.*b’ will still exhibit poor performance.