Rewrite in Vala to fix performance problems (fix #22)
After some experimentation, I found it was hard to fix performance in
Python.

Pre-1.18 versions of rpl were not applying the input regex to more than one
buffer-ful of text, which meant that, for example “rpl 'a*' b foo.txt”,
where foo.txt was a million ‘a’s, would result in a number of b’s, rather
than just one. Having fixed this bug, it transpired that the ‘regex’ library
would run out of memory for fairly modest sizes of input, with regexes like
‘(ab)+’ running out of memory when matching a few million copies.

I tried using PCRE2, with two different bindings (pcre2.py with Cython, and
another with ctypes), but it was much slower than ‘regex’ or the built-in
‘re’ modules. The only remaining choice I could think of was to write my own
Python bindings with the C API, or otherwise delve into the minutiae of
ctypes. I suspected that part of the problem was having to copy strings on
input and output.

So instead, I have rewritten rpl in Vala. At first it seemed that it would
be possible to use GLib’s own regex functions, as they are just a wrapper
around PCRE2, but it turned out not to work, as they use “int” for
positions, so can’t cope with >2GB of data. So, I bound PCRE2.

I was able to use libuchardet for character set detection (and it’s based on
the same Mozilla code as the Python chardet library I was using previously).

PCRE2 itself “only” uses ssize_t for offsets, but my research suggests that
there will not be single disks of this size for some years yet, let alone
files, let alone users who want to process such files in one go.

In the course of doing the translation I found a bug or two, and realised
that the test suite needs considerable improvement. This remains to do.

Also, there’s a certain amount of sub-optimal use of Vala, and I wrote my
own binding to GLib’s Iconv APIs, because I couldn’t make the built-in ones
work. As with past Vala projects, I should be able to make things nicer over
time, and upstream work such as the PCRE2 bindings.