Fix replacement of empty matches
When a match is empty, we advance the search position for the next
iteration by one place to avoid getting stuck, but we must copy all
characters from the end of the previous match into the output, not just
those from where we start the next match!

Thanks to Kevin Maguire for the report.

Add a regression test.