pcre2.vapi: fix possible crash in Regex.substitute() (fix #46)
This method aims to overcome a limitation of the C binding, which can’t
auto-allocate enough memory for the substitution result. There’s no sense in
making the caller deal with this situation. So, ensure that we set
`SUBSTITUTE_OVERFLOW_LENGTH` so that `pcre2_substitute` does indeed return
the amount of memory needed.