Create temporary file in the same directory as file being processed
The temporary file was being created in the current directory! This would
not have worked when the current directory was on a different file system
from the file being processed, as then the temporary file could not be
renamed to the file being processed.

Adjust the test for a non-writable input directory, and remove the test for a
non-writable backup file, as it is no longer possible to detect this
condition, as rpl fails earlier when it can’t create the temporary file.