Make options to substitute the same as to match
PCRE2 10.47 checks when using PCRE2_SUBSITUTE_MATCHED that (inter alia) the
flags are the same for the call to pcre2_substitute as for the pcre2_match
call.

We were adding PCRE2_NOTEMPTY (and I’m not now sure why!). I don’t think it
was having any effect, since we *were* already performing substitutions on
empty matches, so I guess the match flags were being ignored.

We also failed to pass the options in “flags”, which we need to pass, but
without PCRE2_PARTTIAL_HARD, which is not valid for pcre2_substitute.

This still triggers the PCRE 10.47 error, which seems to be a bug.

I have filed https://github.com/PCRE2Project/pcre2/issues/846